LED lights added
Added LED lights using the NEO Pixel Strip. Blue will light up on LED 3 and 4 when the handshake is in progress. When the car is moving forward the LEDs 3 and 4 will display green. While the car is moving backwards the same LEDs will display red. While the car is turning left the LEDS 0 and 1 will display yellow and while the car is turning right the LEDs 6 and 7 will turn yellow.